requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3

# ── Boosting ──────────────────────────────────────────────────────────────────
xgboost>=2.0

# ── Imbalanced ────────────────────────────────────────────────────────────────
imbalanced-learn>=0.11

# ── Explicabilidad ────────────────────────────────────────────────────────────
shap>=0.44

# ── Visualización ─────────────────────────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.12

# ── Data ──────────────────────────────────────────────────────────────────────
pyarrow>=14.0          # parquet
duckdb>=0.10           # SQL sobre parquet en el agente

# ── Agente IA ─────────────────────────────────────────────────────────────────
anthropic>=0.25
python-dotenv>=1.0

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml>=6.0

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0
notebook>=7.0
ipykernel>=6.0
